+++ src/Octoshift/GithubApi.cs
@@ -97,21 +97,6 @@ namespace OctoshiftCLI
             return await _client.GetAllAsync(url).Select(x => (string)x["name"]).ToListAsync();
         }
 
-        public virtual async Task<bool> RepoExists(string org, string repo)
-        {
-            var url = $"{_apiUrl}/repos/{org}/{repo}";
-
-            try
-            {
-                await _client.GetAsync(url);
-                return true;
-            }
-            catch (HttpRequestException ex) when (ex.StatusCode is HttpStatusCode.NotFound or HttpStatusCode.Moved)
-            {
-                return false;
-            }
-        }
-
         public virtual async Task RemoveTeamMember(string org, string teamSlug, string member)
         {
             var url = $"{_apiUrl}/orgs/{org}/teams/{teamSlug}/memberships/{member}";
@@ -176,6 +161,32 @@ namespace OctoshiftCLI
                     url = adoServerUrl,
                     ownerId = orgId,
                     type = "AZURE_DEVOPS"
+                },
+                operationName = "createMigrationSource"
+            };
+
+            var response = await _client.PostAsync(url, payload);
+            var data = JObject.Parse(response);
+
+            return (string)data["data"]["createMigrationSource"]["migrationSource"]["id"];
+        }
+
+        public virtual async Task<string> CreateBbsMigrationSource(string orgId)
+        {
+            var url = $"{_apiUrl}/graphql";
+
+            var query = "mutation createMigrationSource($name: String!, $url: String!, $ownerId: ID!, $type: MigrationSourceType!)";
+            var gql = "createMigrationSource(input: {name: $name, url: $url, ownerId: $ownerId, type: $type}) { migrationSource { id, name, url, type } }";
+
+            var payload = new
+            {
+                query = $"{query} {{ {gql} }}",
+                variables = new
+                {
+                    name = "Bitbucket Server Source",
+                    url = "https://not-used",
+                    ownerId = orgId,
+                    type = "BITBUCKET_SERVER"
                 },
                 operationName = "createMigrationSource"
             };
@@ -328,6 +339,20 @@ namespace OctoshiftCLI
             EnsureSuccessGraphQLResponse(data);
 
             return (string)data["data"]["startOrganizationMigration"]["orgMigration"]["id"];
+        }
+        
+        public virtual async Task<string> StartBbsMigration(string migrationSourceId, string orgId, string repo, string targetToken, string archiveUrl)
+        {
+            return await StartMigration(
+                migrationSourceId,
+                "https://not-used",  // source repository URL
+                orgId,
+                repo,
+                "not-used",  // source access token
+                targetToken,
+                archiveUrl,
+                "https://not-used"  // metadata archive URL
+            );
         }
 
         public virtual async Task<(string State, string RepositoryName, string FailureReason)> GetMigration(string migrationId)