+++ src/OctoshiftCLI.Tests/ado2gh/Commands/MigrateRepoCommandTests.cs
@@ -58,7 +58,6 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
         public async Task Happy_Path()
         {
             // Arrange
-            _mockGithubApi.Setup(x => x.RepoExists(GITHUB_ORG, GITHUB_REPO).Result).Returns(false);
             _mockGithubApi.Setup(x => x.GetOrganizationId(GITHUB_ORG).Result).Returns(GITHUB_ORG_ID);
             _mockGithubApi.Setup(x => x.CreateAdoMigrationSource(GITHUB_ORG_ID, null).Result).Returns(MIGRATION_SOURCE_ID);
             _mockGithubApi
@@ -102,7 +101,6 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
             await _command.Invoke(ADO_ORG, ADO_TEAM_PROJECT, ADO_REPO, GITHUB_ORG, GITHUB_REPO, wait: false);
 
             // Assert
-            _mockGithubApi.Verify(m => m.RepoExists(GITHUB_ORG, GITHUB_REPO));
             _mockGithubApi.Verify(m => m.GetOrganizationId(GITHUB_ORG));
             _mockGithubApi.Verify(m => m.CreateAdoMigrationSource(GITHUB_ORG_ID, null));
             _mockGithubApi.Verify(m => m.StartMigration(MIGRATION_SOURCE_ID, ADO_REPO_URL, GITHUB_ORG_ID, GITHUB_REPO, ADO_TOKEN, GITHUB_TOKEN, null, null, false));
@@ -115,10 +113,24 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
         }
 
         [Fact]
-        public async Task Idempotency_Stop_If_Target_Exists()
+        public async Task Skip_Migration_If_Target_Repo_Exists()
         {
             // Arrange
-            _mockGithubApi.Setup(x => x.RepoExists(GITHUB_ORG, GITHUB_REPO).Result).Returns(true);
+            _mockGithubApi.Setup(x => x.GetOrganizationId(GITHUB_ORG).Result).Returns(GITHUB_ORG_ID);
+            _mockGithubApi.Setup(x => x.CreateAdoMigrationSource(GITHUB_ORG_ID, null).Result).Returns(MIGRATION_SOURCE_ID);
+            _mockGithubApi
+                .Setup(x => x.StartMigration(
+                    MIGRATION_SOURCE_ID,
+                    ADO_REPO_URL,
+                    GITHUB_ORG_ID,
+                    GITHUB_REPO,
+                    ADO_TOKEN,
+                    GITHUB_TOKEN,
+                    null,
+                    null,
+                    false).Result)
+                .Throws(new OctoshiftCliException($"A repository called {GITHUB_ORG}/{GITHUB_REPO} already exists"));
+
             _mockGithubApiFactory.Setup(m => m.Create(It.IsAny<string>(), It.IsAny<string>())).Returns(_mockGithubApi.Object);
 
             _mockEnvironmentVariableProvider
@@ -138,12 +150,8 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
             await _command.Invoke(ADO_ORG, ADO_TEAM_PROJECT, ADO_REPO, GITHUB_ORG, GITHUB_REPO, wait: false);
 
             // Assert
-            _mockGithubApi.Verify(m => m.RepoExists(GITHUB_ORG, GITHUB_REPO));
-
             _mockOctoLogger.Verify(m => m.LogWarning(It.IsAny<string>()), Times.Exactly(1));
             actualLogOutput.Should().Contain(expectedLogOutput);
-
-            _mockGithubApi.VerifyNoOtherCalls();
         }
 
         [Fact]