+++ src/bbs2gh/Commands/MigrateRepoCommand.cs
@@ -14,27 +14,21 @@ namespace OctoshiftCLI.BbsToGithub.Commands
         public MigrateRepoCommand(
             OctoLogger log,
             GithubApiFactory githubApiFactory,
-            EnvironmentVariableProvider environmentVariableProvider) : base("migrate-repo")
+            EnvironmentVariableProvider environmentVariableProvider
+        ) : base("migrate-repo")
         {
             _log = log;
             _githubApiFactory = githubApiFactory;
             _environmentVariableProvider = environmentVariableProvider;
 
-            Description = "Invokes the GitHub API's to migrate the repo and all PR data";
+            Description = "Import a Bitbucket Server archive to GitHub.";
             Description += Environment.NewLine;
-            Description += "Note: Expects ADO_PAT and GH_PAT env variables or --ado-pat and --github-pat options to be set.";
+            Description += "Note: Expects GH_PAT env variable or --github-pat option to be set.";
 
-            var adoOrg = new Option<string>("--ado-org")
+            var archiveUrl = new Option<string>("--archive-url")
             {
-                IsRequired = true
-            };
-            var adoTeamProject = new Option<string>("--ado-team-project")
-            {
-                IsRequired = true
-            };
-            var adoRepo = new Option<string>("--ado-repo")
-            {
-                IsRequired = true
+                IsRequired = true,
+                Description = "URL used to download Bitbucket Server migration archive."
             };
             var githubOrg = new Option<string>("--github-org")
             {
@@ -44,87 +38,74 @@ namespace OctoshiftCLI.BbsToGithub.Commands
             {
                 IsRequired = true
             };
-            var ssh = new Option("--ssh")
-            {
-                IsRequired = false,
-                IsHidden = true,
-                Description = "Uses SSH protocol instead of HTTPS to push a Git repository into the target repository on GitHub."
-            };
-            var wait = new Option("--wait")
-            {
-                IsRequired = false,
-                Description = "Synchronously waits for the repo migration to finish."
-            };
-            var adoPat = new Option<string>("--ado-pat")
-            {
-                IsRequired = false
-            };
             var githubPat = new Option<string>("--github-pat")
             {
                 IsRequired = false
+            };
+            var wait = new Option("--wait")
+            {
+                Description = "Synchronously waits for the repo migration to finish."
             };
             var verbose = new Option("--verbose")
             {
                 IsRequired = false
             };
 
-            AddOption(adoOrg);
-            AddOption(adoTeamProject);
-            AddOption(adoRepo);
+            AddOption(archiveUrl);
             AddOption(githubOrg);
             AddOption(githubRepo);
-            AddOption(ssh);
-            AddOption(wait);
-            AddOption(adoPat);
             AddOption(githubPat);
+            AddOption(wait);
             AddOption(verbose);
 
-            Handler = CommandHandler.Create<string, string, string, string, string, bool, bool, string, string, bool>(Invoke);
+            Handler = CommandHandler.Create<MigrateRepoCommandArgs>(Invoke);
         }
 
-        public async Task Invoke(string adoOrg, string adoTeamProject, string adoRepo, string githubOrg, string githubRepo, bool ssh = false, bool wait = false, string adoPat = null, string githubPat = null, bool verbose = false)
+        public async Task Invoke(MigrateRepoCommandArgs args)
         {
-            _log.Verbose = verbose;
+            if (args is null)
+            {
+                throw new ArgumentNullException(nameof(args));
+            }
+
+            _log.Verbose = args.Verbose;
 
             _log.LogInformation("Migrating Repo...");
-            _log.LogInformation($"ADO ORG: {adoOrg}");
-            _log.LogInformation($"ADO TEAM PROJECT: {adoTeamProject}");
-            _log.LogInformation($"ADO REPO: {adoRepo}");
-            _log.LogInformation($"GITHUB ORG: {githubOrg}");
-            _log.LogInformation($"GITHUB REPO: {githubRepo}");
-            if (ssh)
-            {
-                _log.LogWarning("SSH mode is no longer supported. --ssh flag will be ignored.");
-            }
-            if (wait)
-            {
-                _log.LogInformation("WAIT: true");
-            }
-            if (adoPat is not null)
-            {
-                _log.LogInformation("ADO PAT: ***");
-            }
-            if (githubPat is not null)
+            _log.LogInformation($"GITHUB ORG: {args.GithubOrg}");
+            _log.LogInformation($"GITHUB REPO: {args.GithubRepo}");
+
+            if (args.GithubPat is not null)
             {
                 _log.LogInformation("GITHUB PAT: ***");
             }
-
-            githubPat ??= _environmentVariableProvider.GithubPersonalAccessToken();
-            var githubApi = _githubApiFactory.Create(targetPersonalAccessToken: githubPat);
-            if (await githubApi.RepoExists(githubOrg, githubRepo))
+            if (args.Wait)
             {
-                _log.LogWarning($"The Org '{githubOrg}' already contains a repository with the name '{githubRepo}'. No operation will be performed");
-                return;
+                _log.LogInformation("WAIT: true");
             }
 
-            var adoRepoUrl = GetAdoRepoUrl(adoOrg, adoTeamProject, adoRepo);
+            args.GithubPat ??= _environmentVariableProvider.GithubPersonalAccessToken();
+            var githubApi = _githubApiFactory.Create(targetPersonalAccessToken: args.GithubPat);
+            var githubOrgId = await githubApi.GetOrganizationId(args.GithubOrg);
+            var migrationSourceId = await githubApi.CreateBbsMigrationSource(githubOrgId);
 
-            adoPat ??= _environmentVariableProvider.AdoPersonalAccessToken();
-            var githubOrgId = await githubApi.GetOrganizationId(githubOrg);
-            var migrationSourceId = await githubApi.CreateAdoMigrationSource(githubOrgId, null);
-            var migrationId = await githubApi.StartMigration(migrationSourceId, adoRepoUrl, githubOrgId, githubRepo, adoPat, githubPat);
+            string migrationId;
 
-            if (!wait)
+            try
+            {
+                migrationId = await githubApi.StartBbsMigration(migrationSourceId, githubOrgId, args.GithubRepo, args.GithubPat, args.ArchiveUrl);
+            }
+            catch (OctoshiftCliException ex)
+            {
+                if (ex.Message == $"A repository called {args.GithubOrg}/{args.GithubRepo} already exists")
+                {
+                    _log.LogWarning($"The Org '{args.GithubOrg}' already contains a repository with the name '{args.GithubRepo}'. No operation will be performed");
+                    return;
+                }
+
+                throw;
+            }
+
+            if (!args.Wait)
             {
                 _log.LogInformation($"A repository migration (ID: {migrationId}) was successfully queued.");
                 return;
@@ -147,7 +128,15 @@ namespace OctoshiftCLI.BbsToGithub.Commands
 
             _log.LogSuccess($"Migration completed (ID: {migrationId})! State: {migrationState}");
         }
+    }
 
-        private string GetAdoRepoUrl(string org, string project, string repo) => $"https://dev.azure.com/{org}/{project}/_git/{repo}".Replace(" ", "%20");
+    public class MigrateRepoCommandArgs
+    {
+        public string ArchiveUrl { get; set; }
+        public string GithubOrg { get; set; }
+        public string GithubRepo { get; set; }
+        public string GithubPat { get; set; }
+        public bool Wait { get; set; }
+        public bool Verbose { get; set; }
     }
 }