+++ RELEASENOTES.md
@@ -1,3 +1,1 @@
-- Added `create-team` command to `gh gei`.
-- Changed how we do the version check to avoid warnings about being rate limited
-- Checks for an environment variable GEI_DEBUG_MODE and if set to 'true' will emit additional data in the verbose log file
+- Add `migrate-repo` command to `bbs2gh`.